+++ src/Directory.Packages.props
@@ -51,7 +51,7 @@
     </ItemGroup>
     <PropertyGroup>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <DotNextVersion>5.23.0</DotNextVersion>
+        <DotNextVersion>5.24.0</DotNextVersion>
     </PropertyGroup>
     <PropertyGroup>
         <NoWarn>$(NoWarn);CS1591;CS0535</NoWarn>
@@ -68,7 +68,7 @@
         <PackageVersion Include="DnsClient" Version="1.8.0"/>
         <PackageVersion Include="Newtonsoft.Json" Version="13.0.3"/>
         <PackageVersion Include="Tomat.LibDeflate.Native" Version="1.21.1"/>
-        <PackageVersion Include="System.IO.Pipelines" Version="9.0.7"/>
+        <PackageVersion Include="System.IO.Pipelines" Version="10.0.0-preview.7.25380.108"/>
         <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="8.0.0"/>
         <PackageVersion Include="DotNet.ReproducibleBuilds" Version="1.2.25"/>
     </ItemGroup>
